ref(changelog) : move in a dedicated module
